requirements.txt
absl-py==0.7.1
asgiref==3.2.7
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
boto==2.49.0
boto3==1.12.14
botocore==1.15.14
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cleverhans==3.0.1
cloudpickle==1.1.1
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
Django==3.0.5
dm-sonnet==1.23
docutils==0.15.2
entrypoints==0.3
gast==0.2.2
gensim==3.8.1
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.13.2
grpcio==1.25.0
h5py==2.10.0
idna==2.8
image==1.5.27
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mnist==0.2.2
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
nltk==3.4.5
nose==1.3.7
notebook==6.0.2
numpy==1.17.4
oauthlib==3.1.0
opencv-python==3.4.4.19
opt-einsum==2.3.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.6.0
pydot==1.4.1
Pygments==2.5.2
pyparsing==2.4.5
pyrsistent==0.15.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.21.2
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.13.0
sklearn==0.0
smart-open==1.8.4
sqlparse==0.3.1
tabulate==0.8.5
tensorboard==2.1.1
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
tensorflow-probability==0.8.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
